requirements.txt
playwright>=1.40.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

